+++ lib/nokogiri/html5/document_fragment.rb
@@ -25,6 +25,47 @@ module Nokogiri
     #
     # 💡 HTML5 functionality is not available when running JRuby.
     class DocumentFragment < Nokogiri::HTML4::DocumentFragment
+      class << self
+        # :call-seq:
+        #   parse(tags, **options)
+        #   parse(tags, encoding = nil, **options)
+        #
+        # Parse an HTML5 document fragment from +tags+, returning a Nodeset.
+        #
+        # [Parameters]
+        # - +tags+ [String, IO] The HTML5 document fragment to parse.
+        # - +encoding+ [String] The name of the encoding to use when parsing the document fragment. (default +nil+)
+        #
+        # Also see Nokogiri::HTML5 for a longer explanation of how encoding is handled by the parser.
+        #
+        # [Options]
+        # - +:context+ [String, Nokogiri::XML::Node] The context in which to parse the document fragment. (default +"body"+)
+        # - +:max_errors+ [Integer] The maximum number of parse errors to record. (default +Nokogiri::Gumbo::DEFAULT_MAX_ERRORS+ which is currently 0)
+        # - +:max_tree_depth+ [Integer] The maximum depth of the parse tree. (default +Nokogiri::Gumbo::DEFAULT_MAX_TREE_DEPTH+)
+        # - +:max_attributes+ [Integer] The maximum number of attributes allowed on an element. (default +Nokogiri::Gumbo::DEFAULT_MAX_ATTRIBUTES+)
+        # - +:parse_noscript_content_as_text+ [Boolean] Whether to parse the content of +noscript+ elements as text. (default +false+)
+        #
+        # Also see Nokogiri::HTML5 for a longer explanation of the options.
+        #
+        # [Returns]
+        # - [Nokogiri::XML::NodeSet] A node set containing the root nodes of the parsed fragment.
+        #
+        def parse(tags, encoding = nil, positional_options_hash = nil, **options)
+          unless positional_options_hash.nil?
+            warn("Nokogiri::HTML5::DocumentFragment.parse: Passing options as an explicit hash is deprecated. Use keyword arguments instead. This will become an error in a future release.", uplevel: 1, category: :deprecated)
+            options.merge!(positional_options_hash)
+          end
+
+          context = options.delete(:context)
+
+          document = HTML5::Document.new
+          document.encoding = "UTF-8"
+          tags = HTML5.read_and_encode(tags, encoding)
+
+          new(document, tags, context, options)
+        end
+      end
+
       attr_accessor :document
       attr_accessor :errors
 
@@ -36,32 +77,26 @@ module Nokogiri
       attr_reader :quirks_mode
 
       # Create a document fragment.
-      def initialize(doc, tags = nil, ctx = nil, options = {}) # rubocop:disable Lint/MissingSuper
-        self.document = doc
-        self.errors = []
+      def initialize(doc, tags = nil, context = nil, options = {}) # rubocop:disable Lint/MissingSuper
+        @document = doc
+        @errors = []
         return self unless tags
 
         tags = Nokogiri::HTML5.read_and_encode(tags, nil)
+
+        context = options.delete(:context) if options.key?(:context)
 
         options[:max_attributes] ||= Nokogiri::Gumbo::DEFAULT_MAX_ATTRIBUTES
         options[:max_errors] ||= options.delete(:max_parse_errors) || Nokogiri::Gumbo::DEFAULT_MAX_ERRORS
         options[:max_tree_depth] ||= Nokogiri::Gumbo::DEFAULT_MAX_TREE_DEPTH
 
-        Nokogiri::Gumbo.fragment(self, tags, ctx, **options)
+        Nokogiri::Gumbo.fragment(self, tags, context, **options)
       end
 
       def serialize(options = {}, &block) # :nodoc:
         # Bypass XML::Document.serialize which doesn't support options even
         # though XML::Node.serialize does!
         XML::Node.instance_method(:serialize).bind_call(self, options, &block)
-      end
-
-      # Parse a document fragment from +tags+, returning a Nodeset.
-      def self.parse(tags, encoding = nil, options = {})
-        doc = HTML5::Document.new
-        tags = HTML5.read_and_encode(tags, encoding)
-        doc.encoding = "UTF-8"
-        new(doc, tags, nil, options)
       end
 
       def extract_params(params) # :nodoc: